refactor : Cell, Element, Mine 클래스를 동일한 패키지에 정의
