--- src/main/kotlin/cell/showable/Hide.kt
@@ -1,3 +1,0 @@
-package cell.showable
-
-object Hide : Showable
