+++ src/main/kotlin/element/showable/Hide.kt
@@ -1,0 +1,3 @@
+package element.showable
+
+object Hide : Showable
