+++ src/main/kotlin/element/status/CellStatus.kt
@@ -1,4 +1,4 @@
-package cell.status
+package element.status
 
 interface CellStatus {
     fun isMine(): Boolean = false
