+++ src/main/kotlin/element/status/EmptyCell.kt
@@ -1,3 +1,3 @@
-package cell.status
+package element.status
 
 object EmptyCell : CellStatus
