+++ src/main/kotlin/element/status/MineCell.kt
@@ -1,4 +1,4 @@
-package cell.status
+package element.status
 
 object MineCell : CellStatus {
     override fun isMine(): Boolean = true
