+++ src/main/kotlin/map/Columns.kt
@@ -1,7 +1,7 @@
 package map
 
-import cell.Cell
-import cell.Element
+import element.Cell
+import element.Element
 
 class Columns(
     val points: MutableList<Point>,
